+++ src/Models/Statuses/Status.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Hirtz\Skeleton\Models\Statuses;
 
-use Hirtz\Skeleton\Models\Definitions\Definition;
+use Hirtz\Skeleton\Models\Definitions\ModelDefinition;
 
-class Status extends Definition
+class Status extends ModelDefinition
 {
 }